requirements/local.txt
asgiref==3.2.7
Django==3.0.4
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-model-utils==4.0.0
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
Unipath==1.1